Fix 32 bit warning

+++ src/buffer.cc
@@ -763,7 +763,7 @@ size_t Buffer::read_frame(std::vector<uint8_t> * data, uint32_t frame_no)
 
     data->resize(image_frame->m_size);
 
-    return copy(data->data(), image_frame->m_offset, image_frame->m_size);
+    return copy(data->data(), static_cast<size_t>(image_frame->m_offset), image_frame->m_size);
 }
 
 int Buffer::error() const
